fix: pgtype.Int4 field is Int not Int64 (pgx/v5 API)
Lines 361 and 377 in commandrunner.go: exitCode and durationMs used
Int64 field which does not exist in pgtype.Int4. Correct field is Int
(int32) per pgx/v5 API.

+++ server/internal/handler/commandrunner.go
@@ -358,7 +358,7 @@ func (h *Handler) HandleDaemonCommandRunWS(ctx context.Context, identity daemonw
 
 	var exitCode pgtype.Int4
 	if result.ExitCode >= 0 {
-		exitCode = pgtype.Int4{Int64: int64(result.ExitCode), Valid: true}
+		exitCode = pgtype.Int4{Int: int32(result.ExitCode), Valid: true}
 	}
 
 	now := time.Now()
@@ -374,7 +374,7 @@ func (h *Handler) HandleDaemonCommandRunWS(ctx context.Context, identity daemonw
 
 	var durationMs pgtype.Int4
 	if result.DurationMs >= 0 {
-		durationMs = pgtype.Int4{Int64: int64(result.DurationMs), Valid: true}
+		durationMs = pgtype.Int4{Int: int32(result.DurationMs), Valid: true}
 	}
 
 	_, err = h.Queries.UpdateCommandRunResult(ctx, db.UpdateCommandRunResultParams{
